Simplify headers
