+++ dbms/src/Flash/Coprocessor/GenSchemaAndColumn.h
@@ -18,6 +18,7 @@
 #include <Core/NamesAndTypes.h>
 #include <Flash/Coprocessor/ChunkCodec.h>
 #include <Flash/Coprocessor/TiDBTableScan.h>
+#include <Storages/DeltaMerge/ColumnDefine_fwd.h>
 #include <TiDB/Schema/TiDB.h>
 #include <common/StringRef.h>
 
@@ -31,12 +32,6 @@ NamesAndTypes genNamesAndTypes(const TiDBTableScan & table_scan, const StringRef
 NamesAndTypes genNamesAndTypes(const ColumnInfos & column_infos, const StringRef & column_prefix);
 ColumnsWithTypeAndName getColumnWithTypeAndName(const NamesAndTypes & names_and_types);
 NamesAndTypes toNamesAndTypes(const DAGSchema & dag_schema);
-
-namespace DM
-{
-struct ColumnDefine;
-using ColumnDefinesPtr = std::shared_ptr<std::vector<ColumnDefine>>;
-} // namespace DM
 
 // The column defines and `extra table id index`
 std::tuple<DM::ColumnDefinesPtr, int> genColumnDefinesForDisaggregatedRead(const TiDBTableScan & table_scan);
