+++ dbms/src/Storages/DeltaMerge/Delta/ColumnFileFlushTask.h
@@ -18,7 +18,6 @@
 #include <IO/WriteHelpers.h>
 #include <Storages/DeltaMerge/ColumnFile/ColumnFile.h>
 #include <Storages/DeltaMerge/DeltaIndex.h>
-#include <Storages/DeltaMerge/File/DMFile.h>
 #include <Storages/DeltaMerge/RowKeyRange.h>
 #include <Storages/Page/PageDefinesBase.h>
 #include <common/logger_useful.h>
