+++ dbms/src/Storages/DeltaMerge/Delta/DeltaValueSpace.h
@@ -24,6 +24,7 @@
 #include <Storages/DeltaMerge/ColumnFile/ColumnFileDeleteRange.h>
 #include <Storages/DeltaMerge/ColumnFile/ColumnFileInMemory.h>
 #include <Storages/DeltaMerge/ColumnFile/ColumnFileTiny.h>
+#include <Storages/DeltaMerge/DMContext_fwd.h>
 #include <Storages/DeltaMerge/Delta/ColumnFilePersistedSet.h>
 #include <Storages/DeltaMerge/Delta/MemTableSet.h>
 #include <Storages/DeltaMerge/DeltaIndex.h>
@@ -56,7 +57,6 @@ using DeltaIndexCompacted = DefaultDeltaTree::CompactedEntries;
 using DeltaIndexCompactedPtr = DefaultDeltaTree::CompactedEntriesPtr;
 using DeltaIndexIterator = DeltaIndexCompacted::Iterator;
 
-struct DMContext;
 struct WriteBatches;
 class StoragePool;
