+++ dbms/src/Storages/DeltaMerge/workload/DTWorkload.h
@@ -15,6 +15,7 @@
 #pragma once
 
 #include <Interpreters/Context_fwd.h>
+#include <Storages/DeltaMerge/ColumnDefine_fwd.h>
 #include <fmt/ranges.h>
 
 #include <atomic>
@@ -24,21 +25,13 @@
 
 namespace DB
 {
-class Context;
 class Block;
 } // namespace DB
 
 namespace DB::DM
 {
 class DeltaMergeStore;
-struct ColumnDefine;
-using ColumnDefines = std::vector<ColumnDefine>;
 } // namespace DB::DM
-
-namespace Poco
-{
-class Logger;
-}
 
 namespace DB::DM::tests
 {
