+++ dbms/src/Storages/KVStore/MultiRaft/Disagg/CheckpointIngestInfo.h
@@ -15,6 +15,7 @@
 #pragma once
 
 #include <Common/Logger.h>
+#include <Storages/DeltaMerge/Segment_fwd.h>
 #include <Storages/KVStore/MultiRaft/Disagg/fast_add_peer.pb.h>
 #include <common/types.h>
 
@@ -22,14 +23,6 @@
 
 namespace DB
 {
-namespace DM
-{
-class Segment;
-using SegmentPtr = std::shared_ptr<Segment>;
-using Segments = std::vector<SegmentPtr>;
-struct DMContext;
-using DMContextPtr = std::shared_ptr<DMContext>;
-} // namespace DM
 class Region;
 using RegionPtr = std::shared_ptr<Region>;
 class TMTContext;
